chore: remove the legacy idtable and workspace-record migrations
Both one-time migrations have served their purpose and are removed:

- idtable.MigrateFromKV (core/idtable/migrate.go) — copied a pre-bbolt,
  shared-KV idtable into the standalone bbolt file on first run.
- workspace.MigrateLegacyRecords (internal/core/workspace/migrate.go) —
  converted old "path"-keyed workspace JSON to the new collection.Record format
  before Catalog construction.

Drop both functions, their tests, and the server.go startup calls. The
encodeExtra/decodeExtra helpers that lived in workspace/migrate.go are NOT
migration code (workspace.go's Save, init.go's Init, manage.go, and TestInit use
them to (de)serialize collection.Record.Extra), so they move into workspace.go
next to that Save/Init logic. idtable's LegacyKeyType{Key,NextId} (28/29) are
kept as RESERVED key-space markers (a storage key-collision test still
references them) with an updated comment. The obsolete TestRun_IdTableInitError
(which exercised the idtable migration error path) and TestMigrateLegacyRecords
plus its old-format encoder helpers are removed, and the ~8 test setups that
called MigrateLegacyRecords as a no-op drop the call.

Build, vet, gofmt clean; idtable + workspace + server (and subpkg) suites green
under -race; idtable per-function coverage stays >=84%.
