+++ internal/core/workspace/init_test.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"reflect"
 	"sync"
 	"testing"
 	"time"
@@ -18,7 +17,6 @@ import (
 	"github.com/codetrek/haystack/internal/conf"
 	"github.com/codetrek/haystack/internal/core/storage"
 	"github.com/codetrek/haystack/internal/core/symbols"
-	"github.com/codetrek/haystack/internal/shared/types"
 )
 
 // setupCatalog is a test helper: runs migration, creates collection.Catalog + documents.Store.
@@ -41,13 +39,6 @@ func setupCatalog(t *testing.T, db kv.Store) (cat *collection.Catalog, st *docum
 		idx.CloseAndWait()
 		mpsc.Stop()
 		t.Fatalf("documents.New: %v", err)
-	}
-
-	if err := MigrateLegacyRecords(db, collection.Options{}); err != nil {
-		st.CloseAndWait()
-		idx.CloseAndWait()
-		mpsc.Stop()
-		t.Fatalf("MigrateLegacyRecords: %v", err)
 	}
 
 	cat, err = collection.New(db, st, collection.Options{})
@@ -226,208 +217,6 @@ func TestInitWithMalformedData(t *testing.T) {
 	}
 	if ws.Path != "/valid/path" {
 		t.Errorf("Path mismatch, got %q, want %q", ws.Path, "/valid/path")
-	}
-}
-
-// ----------------------------------------------------------------------------
-// Migration tests
-// ----------------------------------------------------------------------------
-
-// encodeOldWorkspace produces the legacy on-disk JSON that old haystack wrote.
-func encodeOldWorkspace(id int, path string, useGlobal bool, createdAt, lastAccessed, lastFullSync time.Time) []byte {
-	return encodeOldWorkspaceWithFilters(id, path, useGlobal, nil, createdAt, lastAccessed, lastFullSync)
-}
-
-// encodeOldWorkspaceWithFilters is encodeOldWorkspace plus a custom *types.Filters
-// payload, matching the legacy on-disk shape that carried filters inline.
-func encodeOldWorkspaceWithFilters(id int, path string, useGlobal bool, filters *types.Filters, createdAt, lastAccessed, lastFullSync time.Time) []byte {
-	type oldWS struct {
-		ID               int            `json:"id"`
-		Path             string         `json:"path"`
-		UseGlobalFilters bool           `json:"use_global_filters"`
-		Filters          *types.Filters `json:"filters,omitempty"`
-		CreatedAt        time.Time      `json:"created_time"`
-		LastAccessed     time.Time      `json:"last_accessed_time"`
-		LastFullSync     time.Time      `json:"last_full_sync_time"`
-	}
-	b, _ := json.Marshal(oldWS{
-		ID:               id,
-		Path:             path,
-		UseGlobalFilters: useGlobal,
-		Filters:          filters,
-		CreatedAt:        createdAt,
-		LastAccessed:     lastAccessed,
-		LastFullSync:     lastFullSync,
-	})
-	return b
-}
-
-// oldRecordKey encodes the on-disk key for a record using the default prefix.
-func oldRecordKey(id int) []byte {
-	return []byte(fmt.Sprintf("%c%d", collection.DefaultKeyTypeRecord, id))
-}
-
-// TestMigrateLegacyRecords is the required migration proof test.
-//
-// Scenario:
-//   - id=1: OLD format                       → must be migrated to new format.
-//   - id=2: OLD format                       → must be migrated to new format.
-//   - id=3: NEW format                       → must be left untouched.
-//   - id=4: OLD format with non-nil Filters  → filters must round-trip.
-//
-// Assertions after migration:
-//   - collection.New succeeds and sees all four records.
-//   - workspace.Init builds correct *Workspace for each (path, times, UseGlobalFilters, Filters).
-//   - The incr-id counter (key-type 1) is left byte-for-byte unchanged.
-//   - Running migration again is idempotent (on-disk bytes unchanged for id=3,
-//     and count stays at 4).
-func TestMigrateLegacyRecords(t *testing.T) {
-	tempDir := t.TempDir()
-	conf.Get().Global.DataPath = tempDir
-
-	db, err := storage.Open(filepath.Join(tempDir, "data"), 0)
-	if err != nil {
-		t.Fatalf("storage.Open: %v", err)
-	}
-	defer db.Close()
-
-	now := time.Now().UTC().Truncate(time.Second)
-
-	// Seed old-format records.
-	db.Put(oldRecordKey(1), encodeOldWorkspace(1, "/ws/one", true, now, now, now))
-	db.Put(oldRecordKey(2), encodeOldWorkspace(2, "/ws/two", false, now, now, now))
-
-	// id=4: old format carrying a non-nil custom filter set that must survive
-	// migration into the new Extra payload.
-	customFilters := &types.Filters{
-		Include: []string{"*.go", "*.rs"},
-		Exclude: types.Exclude{UseGitIgnore: true, Customized: []string{"target/", "vendor/"}},
-	}
-	db.Put(oldRecordKey(4), encodeOldWorkspaceWithFilters(4, "/ws/four", false, customFilters, now, now, now))
-
-	// Seed a new-format record.
-	newRec3 := collection.Record{
-		ID:           3,
-		Name:         "/ws/three",
-		CreatedAt:    now,
-		LastAccessed: now,
-		Extra:        encodeExtra(true, nil),
-	}
-	newRec3JSON, _ := json.Marshal(newRec3)
-	db.Put(oldRecordKey(3), newRec3JSON)
-
-	// Set the incr-id counter to 4 so future allocations are safe.
-	db.Put([]byte{collection.DefaultKeyTypeIncrId}, []byte("4"))
-
-	// --- Capture incr-id counter + key=3 bytes before any migration ---
-	incrBefore, _ := db.Get([]byte{collection.DefaultKeyTypeIncrId})
-	val3Before, _ := db.Get(oldRecordKey(3))
-
-	// --- Run migration (first time) ---
-	if err := MigrateLegacyRecords(db, collection.Options{}); err != nil {
-		t.Fatalf("MigrateLegacyRecords: %v", err)
-	}
-
-	// --- Verify the incr-id counter (key-type 1) was NOT touched ---
-	incrAfter, _ := db.Get([]byte{collection.DefaultKeyTypeIncrId})
-	if string(incrBefore) != string(incrAfter) {
-		t.Errorf("incr-id counter changed by migration\nbefore: %s\nafter:  %s", incrBefore, incrAfter)
-	}
-
-	// --- Verify key=3 was NOT rewritten by migration ---
-	val3After1, _ := db.Get(oldRecordKey(3))
-	if string(val3Before) != string(val3After1) {
-		t.Errorf("new-format record id=3 was rewritten by migration\nbefore: %s\nafter:  %s", val3Before, val3After1)
-	}
-
-	// --- Construct Catalog + Init ---
-	mpsc := queue.NewMpsc("mig-test-q")
-	mpsc.Start()
-	defer mpsc.Stop()
-
-	st, err := documents.New(db, mpsc, nil, documents.Options{})
-	if err != nil {
-		t.Fatalf("documents.New: %v", err)
-	}
-	defer st.CloseAndWait()
-
-	cat, err := collection.New(db, st, collection.Options{})
-	if err != nil {
-		t.Fatalf("collection.New after migration: %v", err)
-	}
-
-	if err := Init(cat); err != nil {
-		t.Fatalf("Init after migration: %v", err)
-	}
-
-	// --- Assert workspace 1 ---
-	ws1, err := GetByPath("/ws/one")
-	if err != nil {
-		t.Fatalf("GetByPath /ws/one: %v", err)
-	}
-	if ws1.Id != 1 {
-		t.Errorf("ws1.Id = %d, want 1", ws1.Id)
-	}
-	if !ws1.UseGlobalFilters {
-		t.Error("ws1.UseGlobalFilters should be true")
-	}
-	if !ws1.CreatedAt.Equal(now) {
-		t.Errorf("ws1.CreatedAt = %v, want %v", ws1.CreatedAt, now)
-	}
-
-	// --- Assert workspace 2 ---
-	ws2, err := GetByPath("/ws/two")
-	if err != nil {
-		t.Fatalf("GetByPath /ws/two: %v", err)
-	}
-	if ws2.Id != 2 {
-		t.Errorf("ws2.Id = %d, want 2", ws2.Id)
-	}
-	if ws2.UseGlobalFilters {
-		t.Error("ws2.UseGlobalFilters should be false")
-	}
-
-	// --- Assert workspace 3 (already new format) ---
-	ws3, err := GetByPath("/ws/three")
-	if err != nil {
-		t.Fatalf("GetByPath /ws/three: %v", err)
-	}
-	if ws3.Id != 3 {
-		t.Errorf("ws3.Id = %d, want 3", ws3.Id)
-	}
-
-	// --- Assert workspace 4: custom filters round-trip through migration ---
-	ws4, err := GetByPath("/ws/four")
-	if err != nil {
-		t.Fatalf("GetByPath /ws/four: %v", err)
-	}
-	if ws4.Id != 4 {
-		t.Errorf("ws4.Id = %d, want 4", ws4.Id)
-	}
-	if ws4.UseGlobalFilters {
-		t.Error("ws4.UseGlobalFilters should be false")
-	}
-	if ws4.Filters == nil {
-		t.Fatal("ws4.Filters should not be nil after migration")
-	}
-	if !reflect.DeepEqual(ws4.Filters, customFilters) {
-		t.Errorf("ws4.Filters did not round-trip\n got: %+v\nwant: %+v", ws4.Filters, customFilters)
-	}
-
-	// --- Idempotency: running migration again must not change id=3 ---
-	if err := MigrateLegacyRecords(db, collection.Options{}); err != nil {
-		t.Fatalf("second MigrateLegacyRecords: %v", err)
-	}
-
-	val3After2, _ := db.Get(oldRecordKey(3))
-	if string(val3After1) != string(val3After2) {
-		t.Errorf("idempotency failure: id=3 bytes changed on second migration\nbefore: %s\nafter:  %s", val3After1, val3After2)
-	}
-
-	// --- Total count unchanged ---
-	recs := cat.List()
-	if len(recs) != 4 {
-		t.Errorf("catalog.List() = %d records after second migration, want 4", len(recs))
 	}
 }
