+++ internal/server/coverage_test.go
@@ -9,12 +9,10 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/codetrek/haystack/core/idtable"
 	"github.com/codetrek/haystack/core/invertedindex"
 	"github.com/codetrek/haystack/core/kv"
 	"github.com/codetrek/haystack/core/queue"
 	"github.com/codetrek/haystack/internal/conf"
-	"github.com/codetrek/haystack/internal/core/storage"
 	"github.com/codetrek/haystack/internal/shared/running"
 )
 
@@ -102,27 +100,6 @@ func TestRun_IndexStorageError(t *testing.T) {
 	err := run()
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "error initializing index storage")
-}
-
-// TestRun_IdTableInitError tests the run() error path when idtable.New fails
-// (e.g., the data DB contains a corrupt nextId value).
-func TestRun_IdTableInitError(t *testing.T) {
-	tempDir := t.TempDir()
-	conf.Get().Server.CacheSize = 8 * 1024 * 1024
-	conf.Get().Global.DataPath = tempDir
-
-	// Pre-seed a corrupt legacy nextId so the one-time idtable migration fails
-	// (it validates the source counter before copying it into the bbolt file).
-	dataDB, err := storage.Open(filepath.Join(tempDir, "data"), conf.Get().Server.CacheSize)
-	assert.NoError(t, err)
-	// LegacyKeyTypeNextId == 28; write a non-numeric value.
-	err = dataDB.Put([]byte{idtable.LegacyKeyTypeNextId}, []byte("not-a-number"))
-	assert.NoError(t, err)
-	dataDB.Close()
-
-	err = run()
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "error migrating id table")
 }
 
 // TestRun_LockError tests Run() when CheckAndLockServer fails (line 36-38).
